+++ cloud/aws/src/main/java/de/taimos/dvalin/cloud/aws/properties/AWSPropertyProvider.java
@@ -1,0 +1,10 @@
+package de.taimos.dvalin.cloud.aws.properties;
+
+import de.taimos.daemon.properties.BestEffortPropertyProviderChain;
+
+public class AWSPropertyProvider extends BestEffortPropertyProviderChain {
+
+    public AWSPropertyProvider() {
+        this.withProvider(EC2MetadataPropertyProvider.class);
+    }
+}